ncert-offline-rag/requirements.txt
numpy
pandas
faiss-cpu
sentence-transformers
chroma
olama
jsonlines
md5
validators
